+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Upgrade CI deployment google & arcgis auth plugin to v2.0.1
 - #3464: Allow set config for any config fields via helm chart `appConfig` field for all scala services
 - #3466: Allow setting global SQL statement timeout for registry API
+- #3473: Allow to specify service account for cloud sql proxy pod
 
 ## v2.2.4