+++ CHANGES.md
@@ -5,6 +5,8 @@
 - #3463: Fixed Registry API performance issues related to host connection pool & blocking DB IO
 - #3467: Allow to disable scss-compiler job via helm chart config
 - Upgrade CI deployment google & arcgis auth plugin to v2.0.1
+- #3464: Allow set config for any config fields via helm chart `appConfig` field for all scala services
+- #3466: Allow setting global SQL statement timeout for registry API
 - #3473: Allow to specify service account for cloud sql proxy pod
 
 ## v2.2.4